backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.15
django-cors-headers>=4.4
requests>=2.32
python-dotenv>=1.0
